refactor: split ui context menu
